VV-2647 Split the .h file into a .h file and a .cpp file. Remove all the inline specifiers, they are not needed anymore. Make the buffer static and create it once.
